# backend/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Video Retrieval API",
    "description": "API specification for the CLIMB backend, a Video Browser allowing for searching, browsing, and DRES integration (known-item, VQA text answer, and collaborative Ad-hoc Video Search submissions).",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "http://redacted",
      "description": "Online Server"
    }
  ],
  "paths": {
    "/climb/search": {
      "get": {
        "summary": "Search for video segments matching the text query",
        "operationId": "searchVideos",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Text query for search. To exclude specific videos from results, append ' --exclude: videoId1, videoId2, ...' to the query. Example: 'person running --exclude: V3C1_001, V3C1_002'",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number for pagination (default: 1)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "description": "Number of results per page (default: 24, max: 100)",
            "schema": {
              "type": "integer",
              "default": 24,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "avs_session",
            "in": "query",
            "required": false,
            "description": "Active AVS session code. When provided, scenes already submitted in that collaborative session are removed from the results, and the session's idle timer is refreshed.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/climb/dres/connect": {
      "post": {
        "summary": "Authenticate with DRES and fetch the active evaluation ID",
        "operationId": "connectDres",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DresLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connection successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "evaluation_id": {
                      "type": "string"
                    },
                    "selected_name": {
                      "type": "string",
                      "nullable": true
                    },
                    "defaulted": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/climb/dres/submit/kis": {
      "post": {
        "summary": "Submit a video segment to the DRES server (KiS)",
        "operationId": "submitToDres",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DresSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission result, including the verdict reported by DRES",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresSubmissionResult"
                }
              }
            }
          },
          "400": {
            "description": "DRES rejected the submission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresSubmissionError"
                }
              }
            }
          },
          "401": {
            "description": "Not connected to DRES",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresSubmissionError"
                }
              }
            }
          }
        }
      }
    },
    "/climb/dres/submit/vqa": {
      "post": {
        "summary": "Submit a VQA text answer together with the selected media segment to the DRES server",
        "operationId": "submitVqaToDres",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VqaSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission result, including the verdict reported by DRES",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresSubmissionResult"
                }
              }
            }
          },
          "400": {
            "description": "Missing answer text, or DRES rejected the submission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresSubmissionError"
                }
              }
            }
          },
          "401": {
            "description": "Not connected to DRES",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresSubmissionError"
                }
              }
            }
          }
        }
      }
    },
    "/climb/dres/submit/vqa/text": {
      "post": {
        "summary": "Submit a VQA text answer to the DRES server without any media item",
        "operationId": "submitVqaTextOnlyToDres",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VqaTextSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission result, including the verdict reported by DRES",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresSubmissionResult"
                }
              }
            }
          },
          "400": {
            "description": "Missing answer text, or DRES rejected the submission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresSubmissionError"
                }
              }
            }
          },
          "401": {
            "description": "Not connected to DRES",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DresSubmissionError"
                }
              }
            }
          }
        }
      }
    },
    "/climb/avs/session": {
      "post": {
        "summary": "Create a new collaborative AVS session",
        "description": "Creates a fresh session with a random 4-letter (consonant-only) code and returns its snapshot. Sessions are shared by every client of this backend and are deleted after 5 minutes with no interaction (submit, in-session search, or GET poll).",
        "operationId": "createAvsSession",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "nullable": true,
                    "description": "Optional human-readable label for the session"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvsSession"
                }
              }
            }
          }
        }
      }
    },
...
    "/climb/avs/session/{code}/join": {
      "post": {
        "summary": "Join an existing AVS session by code",
        "description": "Returns the session snapshot (case-insensitive code) and refreshes its idle timer. 404 means the session does not exist or has expired.",
        "operationId": "joinAvsSession",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Joined; session snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvsSession"
                }
              }
            }
          },
          "404": {
            "description": "Session not found or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvsSessionError"
                }
              }
            }
          }
        }
      }
    },
    "/climb/avs/session/{code}": {
      "get": {
        "summary": "Fetch an AVS session snapshot (also the poll/heartbeat)",
        "description": "Returns the current scenes, covered videos, and counts for the session, and refreshes its idle timer. Clients poll this endpoint to stay in sync and to keep the session alive; a 404 tells a client its session has expired.",
        "operationId": "getAvsSession",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvsSession"
                }
              }
            }
          },
          "404": {
            "description": "Session not found or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvsSessionError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SearchResult": {
        "type": "object",
        "properties": {
          "video_id": {
            "type": "string"
          },
          "shot_id": {
            "type": "integer"
          },
          "score": {
            "type": "number",
            "format": "float"
          },
          "start_frame": {
            "type": "integer"
          },
          "end_frame": {
            "type": "integer"
          },
          "fps": {
            "type": "number",
            "format": "float"
          },
          "start_time_ms": {
            "type": "integer"
          },
          "end_time_ms": {
            "type": "integer"
          },
          "thumbnail_url": {
            "type": "string"
          }
        },
        "required": [
          "video_id",
          "shot_id",
          "score",
          "start_frame",
          "end_frame",
          "fps",
          "start_time_ms",
          "end_time_ms",
          "thumbnail_url"
        ]
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "per_page": {
            "type": "integer",
            "description": "Number of results per page"
          },
          "count": {
            "type": "integer",
            "description": "Number of results in current page"
          },
          "has_more": {
            "type": "boolean",
            "description": "Whether more results are available beyond the current page"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            }
          }
        },
        "required": [
          "query",
          "page",
          "per_page",
          "count",
          "has_more",
          "results"
        ]
      },
...
      "DresSubmission": {
        "type": "object",
        "description": "Shared by KIS and AVS. The video_id + time range are what DRES scores. For AVS, additionally pass avs_session plus the scene's start_frame/end_frame so the delivered scene is recorded into the shared session and hidden from teammates.",
        "properties": {
          "video_id": {
            "type": "string"
          },
          "start_time_ms": {
            "type": "integer"
          },
          "end_time_ms": {
            "type": "integer"
          },
          "avs_session": {
            "type": "string",
            "nullable": true,
            "description": "AVS session code to record this scene into. Omit for KIS submissions."
          },
          "start_frame": {
            "type": "integer",
            "nullable": true,
            "description": "Scene start frame; identifies the scene (with end_frame) for AVS session recording. A scene spans many keyframes, so recording by frame range hides the whole scene, not just the submitted keyframe."
          },
          "end_frame": {
            "type": "integer",
            "nullable": true,
            "description": "Scene end frame; see start_frame."
          }
        },
        "required": [
          "video_id",
          "start_time_ms",
          "end_time_ms"
        ]
      },
      "VqaSubmission": {
        "type": "object",
        "properties": {
          "text_answer": {
            "type": "string"
          },
          "video_id": {
            "type": "string",
            "nullable": true
          },
          "start_time_ms": {
            "type": "integer",
            "nullable": true
          },
          "end_time_ms": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": [
          "text_answer"
        ]
      },
      "VqaTextSubmission": {
        "type": "object",
        "properties": {
          "text_answer": {
            "type": "string"
          }
        },
        "required": [
          "text_answer"
        ]
      },
      "DresSubmissionResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "pending"
            ],
            "description": "'pending' when DRES accepted the submission but has not graded it yet (HTTP 202)"
          },
          "dres_status": {
            "type": "integer",
            "description": "HTTP status code returned by the DRES submission endpoint"
          },
          "status_summary": {
            "type": "string",
            "description": "Human readable summary of dres_status, e.g. 'Accepted & graded'"
          },
          "verdict": {
            "type": "string",
            "nullable": true,
            "enum": [
              "CORRECT",
              "WRONG",
              "INDETERMINATE",
              "UNDECIDABLE",
              null
            ],
            "description": "The verdict reported by DRES, null while no verdict is available"
          },
          "message": {
            "type": "string"
          },
          "dres_response": {
            "type": "object",
            "description": "The raw SuccessfulSubmissionsStatus payload returned by DRES"
          }
        },
        "required": [
          "status",
          "dres_status",
          "status_summary",
          "message"
        ]
      },
      "DresSubmissionError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "dres_status": {
            "type": "integer",
            "nullable": true,
            "description": "HTTP status code returned by DRES, null when DRES could not be reached"
          },
          "status_summary": {
            "type": "string"
          },
          "details": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "DresLoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "dres_url": {
            "type": "string",
            "default": "https://vbs.videobrowsing.org"
          }
          ,
          "dres_name": {
            "type": "string",
            "description": "Requested evaluation name (e.g. IVADL26)",
            "default": "IVADL26"
          }
        },
        "required": [
          "username",
          "password",
          "dres_url"
        ]
      },
      "AvsScene": {
        "type": "object",
        "description": "A scene submitted within an AVS session, keyed by its frame range.",
        "properties": {
          "video_id": {
            "type": "string"
          },
          "start_frame": {
            "type": "integer"
          },
          "end_frame": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "CORRECT",
              "WRONG",
              "INDETERMINATE",
              "UNDECIDABLE"
            ],
            "description": "DRES verdict at the time the scene was submitted (INDETERMINATE while a live judgement is pending). A CORRECT scene's video is reported in coveredVideos."
          }
        },
        "required": [
          "video_id",
          "start_frame",
          "end_frame",
          "status"
        ]
      },
      "AvsSession": {
        "type": "object",
        "description": "Snapshot of a collaborative AVS session.",
        "properties": {
          "code": {
            "type": "string",
            "description": "4-letter session code"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "scenes": {
            "type": "array",
            "description": "Scenes already submitted in this session (hide these from search / mark them in browse).",
            "items": {
              "$ref": "#/components/schemas/AvsScene"
            }
          },
          "coveredVideos": {
            "type": "array",
            "description": "Video ids with at least one CORRECT hit; dim their remaining shots.",
            "items": {
              "type": "string"
            }
          },
          "counts": {
            "type": "object",
            "properties": {
              "instances": {
                "type": "integer",
                "description": "Number of distinct scenes submitted"
              },
              "distinctVideos": {
                "type": "integer",
                "description": "Number of distinct videos those scenes span"
              }
            }
          },
          "expiresInMs": {
            "type": "integer",
            "description": "Milliseconds until this session idles out, unless an interaction refreshes it"
          }
        },
        "required": [
          "code",
          "scenes",
          "coveredVideos",
          "counts",
          "expiresInMs"
        ]
      },
...
      "AvsSessionError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}